src/bt_app/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bt-app"
version = "0.1.0"
description = "Betaloop application modules for Gazebo camera, MSP, sensors, and trackers."
requires-python = ">=3.10"
dependencies = [
  "loguru>=0.7",
  "numpy>=1.24",
  "opencv-python>=4.8",
  "pyzmq>=25",
]

[project.scripts]
bt-gz-camera-publisher = "bt_app.sensors.gz_camera:main"
bt-image-display = "bt_app.trackers.image_display:main"
bt-red-box-tracker = "bt_app.trackers.red_box_tracker:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["bt_app*"]
